pyproject.toml
[project]
name = "piva-instance-segmentation"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pycocotools (>=2.0.8,<3.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "plotly (>=6.1.2,<7.0.0)",
    "ultralytics (>=8.3.149,<9.0.0)",
    "segmentation-models-pytorch (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

